% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{MCP}
\alias{MCP}
\title{Multiple Correspondence Analysis}
\usage{
MCP(X, k = NULL, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{X}{a data matrix to decompose}

\item{k}{total number of components to return. If \code{NULL}, the full set
of nonzero components are returned.}

\item{tol}{default is \code{.Machine$double.eps^0.5}. A tolerance level for
eliminating effectively zero (small variance) singular value components.}
}
\value{
A list with three elements:
  \item{d}{A vector containing the singular values of X above the
  tolerance threshold.}
  \item{u}{Left (rows) generalized singular vectors. Dimensions are
  \code{nrow(X)} by k.}
  \item{v}{Right (columns) generalized singular vectors. Dimensions are
  \code{ncol(X)} by k.}
}
\description{
The function takes in a data matrix and performs Multiple Correspondence
Analysis (MCP) using Generalized Singular Value Decomposition.
}
