% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_softimpute.R
\name{soft_impute}
\alias{soft_impute}
\alias{generalized_soft_impute}
\title{Soft Imputation}
\usage{
soft_impute(X, lambdas = c(10, 1, 0.1), maxiter = 100, tol = 0.001)

generalized_soft_impute(
  X,
  M,
  W,
  lambdas = c(10, 1, 0.1),
  maxiter = 100,
  tol = 0.001
)
}
\arguments{
\item{X}{matrix}

\item{lambdas}{Tuning parameters}

\item{maxiter}{Maximum number of iterations}

\item{tol}{Tolerance level}

\item{M}{Row constraint matrix}

\item{W}{Column constraint matrix}
}
\description{
Soft Imputation
}
